pyproject.toml
[tool.poetry]
name = "bangdream-gacha-image-generator"
version = "0.1.0"
description = ""
authors = ["Soku_Ritsuki <soku_ritsuki@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = ">=8.3.1"
httpx = ">=0.20.0, <1.0.0"
logoru = "^6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
